Rewrote Java mode in Java.
This was my first foray into using the Scaled APIs from a language that was not
Scala, and boy it sucked. I have some shims in place (SC) to smooth over that
process, but it's not going to be enough.

I'm going to end up writing my own simple collections library (famous last
words re: simple), which I can take the trouble to make comfortable to use from
Scala, Java and eventually other JVM languages. Whee!