Converted to Scaled collections.
This is somewhat anticlimactic, given that all the effort that went into Scaled
collections was so that they would be civilized to use from Java (and other JVM
languages not created by a crazy Swiss guy named Martin). But those dividends
will pay handsomely down the line, not least when I try to make Java mode suck
a bit less.