Revamped to use new Indenter system.
This fixes a few things that weren't working in the old system, and drops
support for blockless if/else if/else and do/while. If you're crazy enough to
want that (without having the body cuddled up next to the control keyword),
then you get to manually maintain that indentation.

The previous system only half-ass supported that anyway, so now we're just
dropping that half of the ass. Providing the whole ass would mean parsing the
Java for reals so that we can actually know when the statement or expression in
a blockless body is actually "done".

That's more than I'm willing to sign up for at the moment.