Just match the java tag.
Minor mode tags are OR not AND, so we don't want to go matching every major
mode which works with projects.

Really what we need to do is ensure that there's a Project available, but since
the java-mode package depends on project-service, that's plus or minus
guaranteed, so it's kind of a non-issue.

I could even just merge this stuff into JavaMode, but I'll cross that bridge
later.