Run unit tests in JDK with sensible version.
If we can tell which version of the JDK a project targets, use that. Otherwise
look for a JDK with the same major version as the "JDK" in which Scaled is
running (the key here being that Scaled may be running in a JRE not a JDK so we
don't want to use thisJDK if we can avoid it).