Tester.isTestFunc -> findTestFunc.
In both cases, look for a properly annotated function, but fall back to the
nearest enclosing function if we don't find one with the right annotation. This
means that we'll sometimes erroneously treat a function as a test function, but
it also means that we'll do our best to do something useful when run against a
language server that doesn't provide any signature information at all. Yay for
compromises.